backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
requests>=2.31.0
aiofiles>=23.2.1
playwright>=1.40.0
pillow>=10.0.0
opencv-python-headless>=4.8.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
